requirements.txt
# requirements.txt
Flask==1.1.2
Flask-Mail==0.9.1
gunicorn==20.0.4
numpy==1.18.5
requests==2.24.0
werkzeug==1.0.1
google-cloud==0.34.0
google-cloud-storage==1.18.0
google-cloud-core == 1.0.0
spleeter==1.5.4
librosa==0.8.0
ffmpeg==1.3
tensorflow==1.15.0
#firebase_admin
